関数をutilに切り出し(表示用ステータス計算とフィールド読み込み)
