+++ src/itemArea/ItemArea.tsx
@@ -10,6 +10,7 @@ import {
     vitaButtonsData, canButtonsData, sealButtonsData, liquidButtonsData,
     hpSpScrollOptions, basicScrollOptions, detailScrollOptions
 } from './itemConfig';
+import { getInputStatus, calculateDisplayStatus } from '../util/StatusUtil';
 
 import { TextField, Box, Grid } from '@mui/material';
 import Button from '@mui/material/Button';
@@ -92,169 +93,11 @@ const ItemArea = (
         setMdfScroll(event.target.value);
     };
 
-    // 入力フィールドを取得する関数
-    const getInputStatus = () => {
-        const newCharacterStatus = { ...characterStatus };
-        newCharacterStatus[STATUS.POW].base = evaluate(inputStatus[FIELDS.CHARA_POW].value).getNumberValue();
-        newCharacterStatus[STATUS.INT].base = evaluate(inputStatus[FIELDS.CHARA_INT].value).getNumberValue();
-        newCharacterStatus[STATUS.VIT].base = evaluate(inputStatus[FIELDS.CHARA_VIT].value).getNumberValue();
-        newCharacterStatus[STATUS.SPD].base = evaluate(inputStatus[FIELDS.CHARA_SPD].value).getNumberValue();
-        newCharacterStatus[STATUS.LUK].base = evaluate(inputStatus[FIELDS.CHARA_LUK].value).getNumberValue();
-        newCharacterStatus[STATUS.POW].card = evaluate(inputStatus[FIELDS.CARD_POW].value).getNumberValue();
-        newCharacterStatus[STATUS.INT].card = evaluate(inputStatus[FIELDS.CARD_INT].value).getNumberValue();
-        newCharacterStatus[STATUS.VIT].card = evaluate(inputStatus[FIELDS.CARD_VIT].value).getNumberValue();
-        newCharacterStatus[STATUS.SPD].card = evaluate(inputStatus[FIELDS.CARD_SPD].value).getNumberValue();
-        newCharacterStatus[STATUS.LUK].card = evaluate(inputStatus[FIELDS.CARD_LUK].value).getNumberValue();
-        newCharacterStatus[STATUS.LEVEL].totalWithoutItem = evaluate(inputStatus[FIELDS.LEVEL].value).getNumberValue();
-        newCharacterStatus[STATUS.HP].totalWithoutItem = evaluate(inputStatus[FIELDS.HP].value).getNumberValue();
-        newCharacterStatus[STATUS.SP].totalWithoutItem = evaluate(inputStatus[FIELDS.SP].value).getNumberValue();
-        newCharacterStatus[STATUS.POW].totalWithoutItem = evaluate(inputStatus[FIELDS.TOTAL_POW].value).getNumberValue();
-        newCharacterStatus[STATUS.INT].totalWithoutItem = evaluate(inputStatus[FIELDS.TOTAL_INT].value).getNumberValue();
-        newCharacterStatus[STATUS.VIT].totalWithoutItem = evaluate(inputStatus[FIELDS.TOTAL_VIT].value).getNumberValue();
-        newCharacterStatus[STATUS.SPD].totalWithoutItem = evaluate(inputStatus[FIELDS.TOTAL_SPD].value).getNumberValue();
-        newCharacterStatus[STATUS.LUK].totalWithoutItem = evaluate(inputStatus[FIELDS.TOTAL_LUK].value).getNumberValue();
-        newCharacterStatus[STATUS.ATK].totalWithoutItem = evaluate(inputStatus[FIELDS.ATK].value).getNumberValue();
-        newCharacterStatus[STATUS.DEF].totalWithoutItem = evaluate(inputStatus[FIELDS.DEF].value).getNumberValue();
-        newCharacterStatus[STATUS.MAT].totalWithoutItem = evaluate(inputStatus[FIELDS.MAT].value).getNumberValue();
-        newCharacterStatus[STATUS.MDF].totalWithoutItem = evaluate(inputStatus[FIELDS.MDF].value).getNumberValue();
-        updateCharacter(newCharacterStatus);
-    };
-
-    // 表示用のステータスを計算する関数
-    const calculateDisplayStatus = () => {
-        const newCharacterStatus = { ...characterStatus };
-        // HPの表示用ステータスを計算
-        newCharacterStatus[STATUS.HP].displayStatus = 
-            newCharacterStatus[STATUS.HP].totalWithoutItem
-            + newCharacterStatus[STATUS.HP].scroll
-            + newCharacterStatus[STATUS.HP].specialSkill;
-        
-        // SPの表示用ステータスを計算
-        newCharacterStatus[STATUS.SP].displayStatus = 
-            newCharacterStatus[STATUS.SP].totalWithoutItem
-            + newCharacterStatus[STATUS.SP].scroll
-            + newCharacterStatus[STATUS.SP].specialSkill;
-        
-        // POWの表示用ステータスを計算
-        newCharacterStatus[STATUS.POW].displayStatus = 
-            newCharacterStatus[STATUS.POW].totalWithoutItem
-            + newCharacterStatus[STATUS.POW].card
-            + newCharacterStatus[STATUS.POW].allVita
-            + newCharacterStatus[STATUS.POW].vita
-            + newCharacterStatus[STATUS.POW].scroll
-            + newCharacterStatus[STATUS.POW].canSeal
-            + newCharacterStatus[STATUS.POW].bradScraper
-            + newCharacterStatus[STATUS.POW].specialSkill;
-        
-        // INTの表示用ステータスを計算
-        newCharacterStatus[STATUS.INT].displayStatus = 
-            newCharacterStatus[STATUS.INT].totalWithoutItem
-            + newCharacterStatus[STATUS.INT].card
-            + newCharacterStatus[STATUS.INT].allVita
-            + newCharacterStatus[STATUS.INT].vita
-            + newCharacterStatus[STATUS.INT].scroll
-            + newCharacterStatus[STATUS.INT].canSeal
-            + newCharacterStatus[STATUS.INT].specialSkill;
-        
-        // SPDの表示用ステータスを計算
-        newCharacterStatus[STATUS.SPD].displayStatus = 
-            newCharacterStatus[STATUS.SPD].totalWithoutItem
-            + newCharacterStatus[STATUS.SPD].card
-            + newCharacterStatus[STATUS.SPD].allVita
-            + newCharacterStatus[STATUS.SPD].vita
-            + newCharacterStatus[STATUS.SPD].scroll
-            + newCharacterStatus[STATUS.SPD].canSeal
-            + newCharacterStatus[STATUS.SPD].specialSkill;
-        
-        // VITの表示用ステータスを計算
-        newCharacterStatus[STATUS.VIT].displayStatus = 
-            newCharacterStatus[STATUS.VIT].totalWithoutItem
-            + newCharacterStatus[STATUS.VIT].card
-            + newCharacterStatus[STATUS.VIT].allVita
-            + newCharacterStatus[STATUS.VIT].vita
-            + newCharacterStatus[STATUS.VIT].scroll
-            + newCharacterStatus[STATUS.VIT].canSeal
-            + newCharacterStatus[STATUS.VIT].specialSkill;
-        
-        // LUKの表示用ステータスを計算
-        newCharacterStatus[STATUS.LUK].displayStatus = 
-            newCharacterStatus[STATUS.LUK].totalWithoutItem
-            + newCharacterStatus[STATUS.LUK].card
-            + newCharacterStatus[STATUS.LUK].allVita
-            + newCharacterStatus[STATUS.LUK].vita
-            + newCharacterStatus[STATUS.LUK].scroll
-            + newCharacterStatus[STATUS.LUK].canSeal
-            + newCharacterStatus[STATUS.LUK].specialSkill;
-        
-        // ATKの表示用ステータスを計算
-        newCharacterStatus[STATUS.ATK].displayStatus = 
-            newCharacterStatus[STATUS.ATK].totalWithoutItem
-            + newCharacterStatus[STATUS.ATK].scroll
-            + newCharacterStatus[STATUS.ATK].specialSkill
-            + newCharacterStatus[STATUS.ATK].liquid
-            + 3*(
-                newCharacterStatus[STATUS.POW].card
-                + newCharacterStatus[STATUS.POW].allVita
-                + newCharacterStatus[STATUS.POW].vita
-                + newCharacterStatus[STATUS.POW].scroll
-                + newCharacterStatus[STATUS.POW].canSeal
-                + newCharacterStatus[STATUS.POW].bradScraper
-                + newCharacterStatus[STATUS.POW].specialSkill
-            )
-            ;
-        
-        // DEFの表示用ステータスを計算
-        newCharacterStatus[STATUS.DEF].displayStatus = 
-            newCharacterStatus[STATUS.DEF].totalWithoutItem
-            + newCharacterStatus[STATUS.DEF].scroll
-            + newCharacterStatus[STATUS.DEF].specialSkill
-            + newCharacterStatus[STATUS.DEF].liquid
-            + 2*(
-                newCharacterStatus[STATUS.VIT].card
-                + newCharacterStatus[STATUS.VIT].allVita
-                + newCharacterStatus[STATUS.VIT].vita
-                + newCharacterStatus[STATUS.VIT].scroll
-                + newCharacterStatus[STATUS.VIT].canSeal
-                + newCharacterStatus[STATUS.VIT].specialSkill
-            );
-        
-        // MATの表示用ステータスを計算
-        newCharacterStatus[STATUS.MAT].displayStatus = 
-            newCharacterStatus[STATUS.MAT].totalWithoutItem
-            + newCharacterStatus[STATUS.MAT].scroll
-            + newCharacterStatus[STATUS.MAT].specialSkill
-            + newCharacterStatus[STATUS.MAT].liquid
-            + 2*(
-                newCharacterStatus[STATUS.INT].card
-                + newCharacterStatus[STATUS.INT].allVita
-                + newCharacterStatus[STATUS.INT].vita
-                + newCharacterStatus[STATUS.INT].scroll
-                + newCharacterStatus[STATUS.INT].canSeal
-                + newCharacterStatus[STATUS.INT].specialSkill
-            );
-        
-        // MDFの表示用ステータスを計算
-        newCharacterStatus[STATUS.MDF].displayStatus = 
-            newCharacterStatus[STATUS.MDF].totalWithoutItem
-            + newCharacterStatus[STATUS.MDF].scroll
-            + newCharacterStatus[STATUS.MDF].specialSkill
-            + newCharacterStatus[STATUS.MDF].liquid
-            + 15*(
-                newCharacterStatus[STATUS.INT].card
-                + newCharacterStatus[STATUS.INT].allVita
-                + newCharacterStatus[STATUS.INT].vita
-                + newCharacterStatus[STATUS.INT].scroll
-                + newCharacterStatus[STATUS.INT].canSeal
-                + newCharacterStatus[STATUS.INT].specialSkill
-            );
-        updateCharacter(newCharacterStatus);
-    };
-
     // ビタボタンクリック時の処理
     const handleVitaButtons: {[key: string]: { handle: () => void; }} = {
         [ITEMS.VITA.ALL.key]: { handle: () => {
             // 入力値を取得
-            getInputStatus();
+            getInputStatus(characterStatus, inputStatus, updateCharacter);
 
             const newCharacterStatus = { ...characterStatus };
             // ビタのステータスを更新
@@ -285,11 +128,11 @@ const ItemArea = (
             };
 
             updateCharacter(newCharacterStatus);
-            calculateDisplayStatus();
+            calculateDisplayStatus(characterStatus, updateCharacter);
         }},
         [ITEMS.VITA.POW.key]: { handle: () => {
             // 入力値を取得
-            getInputStatus();
+            getInputStatus(characterStatus, inputStatus, updateCharacter);
 
             const newCharacterStatus = { ...characterStatus };        
             // ビタのステータスを更新
@@ -299,11 +142,11 @@ const ItemArea = (
                 newCharacterStatus[STATUS.POW].vita = Math.floor(newCharacterStatus[STATUS.POW].base * 0.2);
             };
             updateCharacter(newCharacterStatus);
-            calculateDisplayStatus();
+            calculateDisplayStatus(characterStatus, updateCharacter);
         }},
         [ITEMS.VITA.INT.key]: { handle: () => {
             // 入力値を取得
-            getInputStatus();
+            getInputStatus(characterStatus, inputStatus, updateCharacter);
 
             const newCharacterStatus = { ...characterStatus };
             // ビタのステータスを更新
@@ -313,11 +156,11 @@ const ItemArea = (
                 newCharacterStatus[STATUS.INT].vita = Math.floor(newCharacterStatus[STATUS.INT].base * 0.2);
             };
             updateCharacter(newCharacterStatus);
-            calculateDisplayStatus();
+            calculateDisplayStatus(characterStatus, updateCharacter);
         }},
         [ITEMS.VITA.SPD.key]: { handle: () => {
             // 入力値を取得
-            getInputStatus();
+            getInputStatus(characterStatus, inputStatus, updateCharacter);
 
             const newCharacterStatus = { ...characterStatus };
             // ビタのステータスを更新
@@ -327,11 +170,11 @@ const ItemArea = (
                 newCharacterStatus[STATUS.SPD].vita = Math.floor(newCharacterStatus[STATUS.SPD].base * 0.2);
             }
             updateCharacter(newCharacterStatus);
-            calculateDisplayStatus();
+            calculateDisplayStatus(characterStatus, updateCharacter);
         }},
         [ITEMS.VITA.VIT.key]: { handle: () => {
             // 入力値を取得
-            getInputStatus();
+            getInputStatus(characterStatus, inputStatus, updateCharacter);
 
             const newCharacterStatus = { ...characterStatus };
             // ビタのステータスを更新
@@ -341,11 +184,11 @@ const ItemArea = (
                 newCharacterStatus[STATUS.VIT].vita = Math.floor(newCharacterStatus[STATUS.VIT].base * 0.2);
             }
             updateCharacter(newCharacterStatus);
-            calculateDisplayStatus();
+            calculateDisplayStatus(characterStatus, updateCharacter);
         }},
         [ITEMS.VITA.LUK.key]: { handle: () => {
             // 入力値を取得
-            getInputStatus();
+            getInputStatus(characterStatus, inputStatus, updateCharacter);
 
             const newCharacterStatus = { ...characterStatus };
             // ビタのステータスを更新
@@ -355,7 +198,7 @@ const ItemArea = (
                 newCharacterStatus[STATUS.LUK].vita = Math.floor(newCharacterStatus[STATUS.LUK].base * 0.2);
             }
             updateCharacter(newCharacterStatus);
-            calculateDisplayStatus();
+            calculateDisplayStatus(characterStatus, updateCharacter);
         }},
     };
 
@@ -363,7 +206,7 @@ const ItemArea = (
     const handleCanButtons: {[key: string]: { handle: () => void; }} = {
         [ITEMS.CAN.A.key]: { handle: () => {
             // 入力値を取得
-            getInputStatus();
+            getInputStatus(characterStatus, inputStatus, updateCharacter);
 
             const newCharacterStatus = { ...characterStatus };
 
@@ -371,7 +214,7 @@ const ItemArea = (
         }},
         [ITEMS.CAN.B.key]: { handle: () => {
             // 入力値を取得
-            getInputStatus();
+            getInputStatus(characterStatus, inputStatus, updateCharacter);
 
             const newCharacterStatus = { ...characterStatus };
